+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.44.0] - 2021-04-12
+
+### Added
+- DropDown: Added class disabled
+
+### Fixed
+- NumericTextBox: Fixed rounding numbers on change event
+
+
 ## [0.43.0] - 2021-03-29
 
 ### Added